+++ appnav/src/main/kotlin/io/element/android/appnav/BackstackExt.kt
@@ -16,12 +16,6 @@
 
 package io.element.android.appnav
 
-import android.content.Context
-import android.content.ContextWrapper
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.BackHandler
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalContext
 import com.bumble.appyx.navmodel.backstack.BackStack
 import com.bumble.appyx.navmodel.backstack.operation.NewRoot
 import com.bumble.appyx.navmodel.backstack.operation.Remove
@@ -46,19 +40,3 @@ fun <T : Any> BackStack<T>.removeLast(element: T) {
     } ?: return
     accept(Remove(lastExpectedNavElement.key))
 }
-
-@Composable
-fun MoveActivityToBackgroundBackHandler(enabled: Boolean = true) {
-
-    fun Context.findActivity(): ComponentActivity? = when (this) {
-        is ComponentActivity -> this
-        is ContextWrapper -> baseContext.findActivity()
-        else -> null
-    }
-
-    val context = LocalContext.current
-    BackHandler(enabled = enabled) {
-        context.findActivity()?.moveTaskToBack(false)
-    }
-}
-
